Rename json_encoder.py to encoder.py
